requirements.txt
bayesian-optimization==2.0.3
catboost==1.2.7
chinesecalendar==1.10.0
et_xmlfile==2.0.0
executing==2.2.0
fastexcel==0.13.0
fastjsonschema==2.21.1
graphviz==0.20.3
lightgbm==4.6.0
loguru==0.7.3
matplotlib==3.10.1
multiprocess==0.70.18
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
optuna==4.2.1
pandas==2.2.3
scikit-learn==1.4.2
scipy==1.15.2
seaborn==0.13.2
shap==0.47.1
sklearn-genetic==0.6.0
statsmodels==0.14.4
tqdm==4.66.5
workalendar==17.0.0
xgboost==2.0.0